完成部分页面

+++ server/main/src/main/java/com/smileshark/entity/Customer.java
@@ -34,7 +34,7 @@ public class Customer {
      * 积分
      */
     @TableField("integral")
-    private Long integral;
+    private Integer integral;
 
     /**
      * 客户对用的钱包
